pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [
  {name = "Yuxuan Lu", email = "luyuxuanleo@gmail.com"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "aioboto3>=14.3.0",
    "anthropic>=0.69.0",
    "beautifulsoup4>=4.13.4",
    "boto3>=1.37.3",
    "click>=8.1.8",
    "dominate>=2.9.1",
    "flask>=3.0.3",
    "gymnasium==0.29.1",
    "hydra-core>=1.3.2",
    "ipywidgets>=8.1.7",
    "json-fix>=1.0.0",
    "jupyterlab>=4.3.7",
    "litellm>=1.77.7",
    "omegaconf>=2.3.0",
    "openai>=1.86.0",
    "pandas>=2.0.3",
    "playwright>=1.48.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pytz>=2025.2",
    "selenium>=4.27.1",
    "tiktoken>=0.7.0",
    "tqdm>=4.67.1",
    "werkzeug>=3.0.6",
]
description = ''
dynamic = ["version"]
keywords = []
license = "MIT"
name = "simulated-web-agent"
readme = "README.md"
requires-python = ">=3.8"

[project.urls]
Documentation = "https://github.com/Yuxuan Lu/simulated-web-agent#readme"
Issues = "https://github.com/Yuxuan Lu/simulated-web-agent/issues"
Source = "https://github.com/Yuxuan Lu/simulated-web-agent"

[tool.hatch.version]
path = "src/simulated_web_agent/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/simulated_web_agent tests}"

[tool.coverage.run]
branch = true
omit = [
  "src/simulated_web_agent/__about__.py",
]
parallel = true
source_pkgs = ["simulated_web_agent", "tests"]

[tool.coverage.paths]
simulated_web_agent = ["src/simulated_web_agent", "*/simulated-web-agent/src/simulated_web_agent"]
tests = ["tests", "*/simulated-web-agent/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.ruff.lint]
ignore = ["F401"]
select = ["E4", "E7", "E9", "F", "I"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.basedpyright]
typeCheckingMode = "basic"
